requirements.txt
astroid==2.4.2
autopep8==1.5.4
isort==5.6.4
kafka-python==2.0.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
six==1.15.0
toml==0.10.2
wrapt==1.12.1
